+++ templates/service-express/server/src/entities/config/system.config.ts
@@ -30,45 +30,6 @@ interface IFileLogConf extends ILogConf {
     readonly rotate_days?: number;
 }
 
-//#module graylog
-interface IGraylogConf extends ILogConf {
-    /**
-     * 服务器列表
-     */
-    readonly servers: {
-        /**
-         * 服务器地址
-         */
-        readonly host: string,
-
-        /**
-         * 端口
-         * @type integer
-         */
-        readonly port: number
-    }[];
-}
-//#endmodule graylog
-
-//#module loki
-interface ILokiConf extends ILogConf {
-    readonly host: string;
-
-    /**
-     * 日志发送时间间隔 (秒)
-     * @type integer
-     * @default 5
-     */
-    readonly interval?: number;
-
-    /**
-     * 是否批量发送日志
-     * @default true
-     */
-    readonly batching?: boolean;
-}
-//#endmodule loki
-
 //#module sentry
 interface ISentryConf extends ILogConf {
     readonly dsn: string;
@@ -93,12 +54,6 @@ export interface ISystemConf {
     readonly log: {
         readonly console?: IConsoleLogConf,
         readonly file?: IFileLogConf,
-        //#module graylog
-        readonly graylog?: IGraylogConf,
-        //#endmodule graylog
-        //#module loki
-        readonly loki?: ILokiConf,
-        //#endmodule loki
         //#module sentry
         readonly sentry?: ISentryConf,
         //#endmodule sentry